+++ test/views/subviewWithEvents.js
@@ -1,0 +1,8 @@
+define(['backbone'], function(Backbone){
+  return Backbone.View.extend({
+    className: 'sub-view',
+    events: {
+      click: function(){ this.$el.html('clicked'); }
+    }
+  });
+});