+++ test/views/subviewWithModel.js
@@ -1,0 +1,8 @@
+define(['backbone'], function(Backbone){
+  return Backbone.View.extend({
+    className: 'sub-view',
+    render: function(){
+      this.$el.html(this.model);
+    }
+  });
+});